feat(emergency): add new routes and tab pages
- Split emergency routes to a new file, update casing of route files
- Add routes, placeholder views for OperationStrategy, ActionsSummary, Background
