+++ app/src/App/routes/index.tsx
@@ -11,10 +11,11 @@ import {
     customWrapRoute,
     rootLayout,
 } from './common';
-import countryRoutes from './CountryRoutes';
-import regionRoutes from './RegionRoutes';
+import countryRoutes from './countryRoutes';
+import emergencyRoutes from './emergencyRoutes';
+import regionRoutes from './regionRoutes';
 import SmartNavigate from './SmartNavigate';
-import surgeRoutes from './SurgeRoutes';
+import surgeRoutes from './surgeRoutes';
 
 const fourHundredFour = customWrapRoute({
     parent: rootLayout,
@@ -113,19 +114,6 @@ const home = customWrapRoute({
     },
 });
 
-const emergencies = customWrapRoute({
-    parent: rootLayout,
-    path: 'emergencies',
-    component: {
-        render: () => import('#views/Emergencies'),
-        props: {},
-    },
-    wrapperComponent: Auth,
-    context: {
-        title: 'Emergencies',
-        visibility: 'anything',
-    },
-});
 const cookiePolicy = customWrapRoute({
     parent: rootLayout,
     path: 'cookie-policy',
@@ -136,185 +124,6 @@ const cookiePolicy = customWrapRoute({
     wrapperComponent: Auth,
     context: {
         title: 'Cookie Policy',
-        visibility: 'anything',
-    },
-});
-
-type DefaultEmergenciesChild = 'details';
-const emergenciesLayout = customWrapRoute({
-    parent: rootLayout,
-    path: 'emergencies/:emergencyId',
-    forwardPath: 'details' satisfies DefaultEmergenciesChild,
-    component: {
-        render: () => import('#views/Emergency'),
-        props: {},
-    },
-    wrapperComponent: Auth,
-    context: {
-        title: 'Emergency',
-        visibility: 'anything',
-    },
-});
-
-const emergencySlug = customWrapRoute({
-    parent: rootLayout,
-    path: 'emergencies/slug/:slug',
-    component: {
-        render: () => import('#views/EmergencySlug'),
-        props: {},
-    },
-    wrapperComponent: Auth,
-    context: {
-        title: 'Emergency',
-        visibility: 'anything',
-    },
-});
-
-const emergencyFollow = customWrapRoute({
-    parent: rootLayout,
-    path: 'emergencies/:emergencyId/follow',
-    component: {
-        render: () => import('#views/EmergencyFollow'),
-        props: {},
-    },
-    wrapperComponent: Auth,
-    context: {
-        title: 'Follow Emergency',
-        visibility: 'is-authenticated',
-    },
-});
-
-const emergencyIndex = customWrapRoute({
-    parent: emergenciesLayout,
-    index: true,
-    component: {
-        eagerLoad: true,
-        render: SmartNavigate,
-        props: {
-            to: 'details' satisfies DefaultEmergenciesChild,
-            replace: true,
-            hashToRouteMap: {
-                '#details': 'details',
-                '#reports': 'reports',
-                '#activities': 'activities',
-                '#surge': 'surge',
-            },
-            // TODO: make this typesafe
-            forwardUnmatchedHashTo: 'additional-info',
-        },
-    },
-    context: {
-        title: 'Emergency',
-        visibility: 'anything',
-    },
-});
-
-const emergencyDetails = customWrapRoute({
-    parent: emergenciesLayout,
-    path: 'details' satisfies DefaultEmergenciesChild,
-    component: {
-        render: () => import('#views/EmergencyDetails'),
-        props: {},
-    },
-    context: {
-        title: 'Emergency Details',
-        visibility: 'anything',
-    },
-});
-
-const emergencyReportsAndDocuments = customWrapRoute({
-    parent: emergenciesLayout,
-    path: 'reports',
-    component: {
-        render: () => import('#views/EmergencyReportAndDocument'),
-        props: {},
-    },
-    context: {
-        title: 'Emergency Reports and Documents',
-        visibility: 'anything',
-    },
-});
-
-const emergencyActivities = customWrapRoute({
-    parent: emergenciesLayout,
-    path: 'activities',
-    component: {
-        render: () => import('#views/EmergencyActivities'),
-        props: {},
-    },
-    context: {
-        title: 'Emergency Activities',
-        visibility: 'anything',
-    },
-});
-const emergencySurge = customWrapRoute({
-    parent: emergenciesLayout,
-    path: 'surge',
-    component: {
-        render: () => import('#views/EmergencySurge'),
-        props: {},
-    },
-    context: {
-        title: 'Emergency Surge',
-        visibility: 'anything',
-    },
-});
-
-// TODO: remove this route
-const emergencyAdditionalInfoOne = customWrapRoute({
-    parent: emergenciesLayout,
-    path: 'additional-info-1',
-    component: {
-        render: () => import('#views/EmergencyAdditionalTab'),
-        props: {
-            infoPageId: 1,
-        },
-    },
-    context: {
-        title: 'Emergency Additional Tab 1',
-        visibility: 'anything',
-    },
-});
-// TODO: remove this route
-const emergencyAdditionalInfoTwo = customWrapRoute({
-    parent: emergenciesLayout,
-    path: 'additional-info-2',
-    component: {
-        render: () => import('#views/EmergencyAdditionalTab'),
-        props: {
-            infoPageId: 2,
-        },
-    },
-    context: {
-        title: 'Emergency Additional Tab 2',
-        visibility: 'anything',
-    },
-});
-// TODO: remove this route
-const emergencyAdditionalInfoThree = customWrapRoute({
-    parent: emergenciesLayout,
-    path: 'additional-info-3',
-    component: {
-        render: () => import('#views/EmergencyAdditionalTab'),
-        props: {
-            infoPageId: 3,
-        },
-    },
-    context: {
-        title: 'Emergency Additional Tab 3',
-        visibility: 'anything',
-    },
-});
-
-const emergencyAdditionalInfo = customWrapRoute({
-    parent: emergenciesLayout,
-    path: 'additional-info/:tabId?',
-    component: {
-        render: () => import('#views/EmergencyAdditionalTab'),
-        props: {},
-    },
-    context: {
-        title: 'Emergency Additional Info Tab',
         visibility: 'anything',
     },
 });
@@ -1455,20 +1264,7 @@ const wrappedRoutes = {
     recoverAccountConfirm,
     resendValidationEmail,
     home,
-    emergencies,
     cookiePolicy,
-    emergencySlug,
-    emergencyFollow,
-    emergenciesLayout,
-    emergencyDetails,
-    emergencyIndex,
-    emergencyReportsAndDocuments,
-    emergencyActivities,
-    emergencySurge,
-    emergencyAdditionalInfoOne,
-    emergencyAdditionalInfoTwo,
-    emergencyAdditionalInfoThree,
-    emergencyAdditionalInfo,
     preparednessLayout,
     preparednessGlobalSummary,
     preparednessGlobalPerformance,
@@ -1534,6 +1330,7 @@ const wrappedRoutes = {
     ...regionRoutes,
     ...countryRoutes,
     ...surgeRoutes,
+    ...emergencyRoutes,
 
     // TODO: Remove me after implementation of DrefFinalReport for imminent
     oldDrefFinalReportForm,
