+++ app/src/views/Emergency/i18n.json
@@ -4,6 +4,9 @@
         "emergencyPageTitle": "IFRC GO - Emergency - {emergencyName}",
         "emergencyPageTitleFallback": "IFRC GO - Emergency",
         "emergencyTabDetails":"Emergency Details",
+        "emergencyTabActionsSummary": "Actions Summary",
+        "emergencyTabOperationStrategy": "Operation Strategy",
+        "emergencyTabBackground": "Background",
         "emergencyTabReports":"Reports/Documents",
         "emergencyTabActivities":"Activities",
         "emergencyTabSurge":"Surge",
