+++ app/src/views/EmergencyActionsSummary/index.tsx
@@ -1,0 +1,13 @@
+import TabPage from '#components/TabPage';
+
+/** @knipignore */
+// eslint-disable-next-line import/prefer-default-export
+export function Component() {
+    return (
+        <TabPage>
+            Emergency Actions Summary
+        </TabPage>
+    );
+}
+
+Component.displayName = 'EmergencyActions Summary';
