+++ app/src/views/EmergencyOperationStrategy/index.tsx
@@ -1,0 +1,13 @@
+import TabPage from '#components/TabPage';
+
+/** @knipignore */
+// eslint-disable-next-line import/prefer-default-export
+export function Component() {
+    return (
+        <TabPage>
+            Emergency Operation Strategy
+        </TabPage>
+    );
+}
+
+Component.displayName = 'EmergencyOperationStrategy';
